requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
ibm-watsonx-ai
python-dotenv
numpy